move importer outside of entries namespace
